refactor: 整理测试文件和 cmd 工具
测试文件合并 (25→16):
- convert: 10→5 (碎片合并到 converter_test/equivalence/e2e)
- templates: 6→4 (执行测试合并为 impl_test, compile→template 重命名)
- operators: 3→2 (convert_integration 合入 convert_test)
- protocols/http: 4→3 (favicon_e2e 合入 request_test)
- common/tlsx: 2→1 (cipher 合入 tlsx_test)

cmd 工具整理 (8→4):
- shot: 用 templates.Load 替代手动 yaml 解析,支持 xray 自动转换
- validate: 吸收 pocverify 为 compare 子命令
- xrayconv → converter: 与 xray 语义脱绑
- 删除: demo_reqresp, runpoc, pocverify
