+++ cmd/validate/validate.go
@@ -1,3 +1,10 @@
+// validate checks that neutron/nuclei templates compile correctly, and
+// optionally compares xray POC semantics against converted templates.
+//
+// Usage:
+//
+//	validate <path_or_file>                                     # compile check
+//	validate compare --xray <xray.yml> --nuclei <nuclei.yaml>   # semantic comparison
 package main
 
 import (
@@ -6,45 +13,50 @@ import (
 	"os"
 	"path/filepath"
 
+	_ "github.com/chainreactors/neutron/convert"
 	"github.com/chainreactors/neutron/protocols"
 	"github.com/chainreactors/neutron/templates"
-	"gopkg.in/yaml.v3"
 )
-
-var ExecuterOptions *protocols.ExecuterOptions
 
 func main() {
 	if len(os.Args) < 2 {
 		fmt.Println("Usage: validate <path_or_file>")
+		fmt.Println("       validate compare --xray <xray.yml> --nuclei <nuclei.yaml> [--json] [-v]")
+		os.Exit(1)
+	}
+
+	if os.Args[1] == "compare" {
+		runCompare(os.Args[2:])
 		return
 	}
 
 	target := os.Args[1]
+	opts := &protocols.ExecuterOptions{Options: &protocols.Options{Timeout: 5}}
 	err := filepath.Walk(target, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			return err
 		}
-		if filepath.Ext(path) == ".yaml" || filepath.Ext(path) == ".yml" {
-			content, err := ioutil.ReadFile(path)
-			if err != nil {
-				return err
-			}
-
-			t := &templates.Template{}
-			err = yaml.Unmarshal(content, &t)
-			if err != nil {
-				fmt.Printf("FAIL %s - YAML unmarshalling failed: %s\n", path, err.Error())
-				return nil
-			}
-
-			err = t.Compile(ExecuterOptions)
-			if err != nil {
-				fmt.Printf("FAIL %s - Template compilation failed: %s\n", path, err.Error())
-				return nil
-			}
-
-			fmt.Printf("OK   %s\n", path)
+		if filepath.Ext(path) != ".yaml" && filepath.Ext(path) != ".yml" {
+			return nil
 		}
+		content, err := ioutil.ReadFile(path)
+		if err != nil {
+			return err
+		}
+
+		t, err := templates.Load(content)
+		if err != nil {
+			fmt.Printf("FAIL %s - load failed: %s\n", path, err.Error())
+			return nil
+		}
+
+		err = t.Compile(opts)
+		if err != nil {
+			fmt.Printf("FAIL %s - compile failed: %s\n", path, err.Error())
+			return nil
+		}
+
+		fmt.Printf("OK   %s\n", path)
 		return nil
 	})
